Aggregate of HDRP bugfixes
[Fix Texture Field to accept only 2D Textures](https://github.cds.internal.unity3d.com/unity/unity/pull/8549)

[Fixed the documentation for recursive rendering not being clear enough for the smoothness' behavior ](https://github.cds.internal.unity3d.com/unity/unity/pull/8643)

[[HDRP][Path Tracing] Making sure there is no unwanted camera jittering](https://github.cds.internal.unity3d.com/unity/unity/pull/9419)

[Update HDR in comparison docs #9086](https://github.cds.internal.unity3d.com/unity/unity/pull/9086)

[[HDRP][Path Tracing] Facilitated support for Ray Tracing Light Cluster in Path Tracing](https://github.cds.internal.unity3d.com/unity/unity/pull/9064)

[Update thick transmission docs](https://github.cds.internal.unity3d.com/unity/unity/pull/8568)

[Remove check for TAA when checking if pyramid history is required.](https://github.cds.internal.unity3d.com/unity/unity/pull/8746)

[[HDRP] Fix cloud layer overflow due to alpha blending](https://github.cds.internal.unity3d.com/unity/unity/pull/8565)

[Force sync compilation on SMAA shaders](https://github.cds.internal.unity3d.com/unity/unity/pull/8555)

[Fix Light Loop Performance Issue with HDRP Rendering Debugger](https://github.cds.internal.unity3d.com/unity/unity/pull/8496)

[Fix Decal Scaling](https://github.cds.internal.unity3d.com/unity/unity/pull/8464)

[Allow for null parameter for Texture2D and Texture3D parameters](https://github.cds.internal.unity3d.com/unity/unity/pull/8463)

[Fix scalarization issues on gamecore](https://github.cds.internal.unity3d.com/unity/unity/pull/8433)

[Fixed the ambient probe for the volumetric clouds for the sky cubemap not being ready at the first frame (case 1403468). ](https://github.cds.internal.unity3d.com/unity/unity/pull/8650)

[Fix two volumetric clouds issues on HDRP](https://github.cds.internal.unity3d.com/unity/unity/pull/8636)

[Fixed the SSR not properly working on deferred with tiles with multiple variants (case 1399311).](https://github.cds.internal.unity3d.com/unity/unity/pull/8533)

[Fixed artifacts on the volumetric clouds when enabling the fog (case 1407759)](https://github.cds.internal.unity3d.com/unity/unity/pull/8503)

[Fixed the HDRP RTAS Debug view tests not working (case 1404219).](https://github.cds.internal.unity3d.com/unity/unity/pull/8459)

+++ Packages/com.unity.render-pipelines.high-definition/Runtime/Lighting/VolumetricLighting/VolumetricCloudsDef.cs
@@ -149,9 +149,9 @@ namespace UnityEngine.Rendering.HighDefinition
         // Flag that allows us to know if the maxZMask texture is valid
         public int _ValidMaxZMask;
         // Flag that allows to know if we should be using the improved transmittance blending
-        public int _ImprovedTransmittanceBlend;
+        public float _ImprovedTransmittanceBlend;
         // Flag that defines if the transmittance should follow a cubic profile (For MSAA)
-        public int _CubicTransmittance;
+        public float _CubicTransmittance;
         public int _Padding1;
 
         [HLSLArray(3 * 4, typeof(Vector4))]
