+++ Packages/com.unity.render-pipelines.high-definition/Runtime/Lighting/VolumetricLighting/VolumetricCloudsDef.cs.hlsl
@@ -70,8 +70,8 @@ CBUFFER_START(ShaderVariablesClouds)
     int _EnableFastToneMapping;
     int _IsPlanarReflection;
     int _ValidMaxZMask;
-    int _ImprovedTransmittanceBlend;
-    int _CubicTransmittance;
+    float _ImprovedTransmittanceBlend;
+    float _CubicTransmittance;
     int _Padding1;
     float4 _DistanceBasedWeights[12];
 CBUFFER_END
